% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_denoise.R
\name{benchmark_denoise}
\alias{benchmark_denoise}
\title{Benchmark denoising on first channel with Gaussian noise}
\usage{
benchmark_denoise(
  img,
  sigma = 0.05,
  iterations = 500,
  minlossimprove = 1e-04,
  smband = 10
)
}
\arguments{
\item{img}{input image (matrix/array, or integer specifying size)}

\item{sigma}{noise standard deviation (in 0–1 scale)}

\item{iterations}{number of noisy realizations (default 500)}

\item{minlossimprove}{parameter for smoothimage}

\item{smband}{parameter for smoothimage}
}
\value{
list with mean_rmse, sd_rmse, all_rmse
}
\description{
Benchmark denoising on first channel with Gaussian noise
}
