% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcpp.R
\name{demo2d}
\alias{demo2d}
\title{Title demo2d
Demonstrates denoising performance on simple 2d test image}
\usage{
demo2d(size = 256, sigma = 0.1, smband = 10, tolmult = 0.001)
}
\arguments{
\item{size}{resolution of the test image to be taken}

\item{sigma}{standard deviation of the noise to add}

\item{smband}{smoothing band}

\item{tolmult}{Multiplier to calculate cutoff for branching}
}
\description{
Title demo2d
Demonstrates denoising performance on simple 2d test image
}
\examples{
demo2d(size=256,sigma=0.1,smband=10,tolmult=0.001)
}
