% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcpp.R
\name{demo3d}
\alias{demo3d}
\title{Title demo3d
Demo denoising on 3D images in 0-1 scale}
\usage{
demo3d(
  size = c(128, 128, 128),
  sigma = 0.1,
  smband = 4,
  tolmult = 0.01,
  margin = 0.05,
  slice_idx = NULL,
  invert = FALSE
)
}
\arguments{
\item{size}{dimensions of the image}

\item{sigma}{standard deviation of the noise to add}

\item{smband}{smoothing band}

\item{tolmult}{multiplier of tolerance, default is set to 0.01, determines the cutoff for branching}

\item{margin}{space to leave between the demo image and boundary}

\item{slice_idx}{slices to plot}

\item{invert}{Invert image color}
}
\value{
list of noisy image and smoothed image
}
\description{
Title demo3d
Demo denoising on 3D images in 0-1 scale
}
\examples{
demo3d(size=c(128,128,128),sigma=0.1,smband=4,tolmult=0.01,margin=0.05,slice_idx=NULL,invert=FALSE)
}
