% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testcpp.R
\name{demo3diter}
\alias{demo3diter}
\title{Title demo3diter
Multiple iterations of 3D demo}
\usage{
demo3diter(
  image = 128,
  sigma = 0.1,
  smband = 4,
  tolmult = 0.01,
  margin = 0.05,
  slice_idx = NULL,
  invert = FALSE,
  n_iter = 10,
  plot_instance = 1
)
}
\arguments{
\item{image}{Integer value represents dimensions of the images, location of a .nii file can be used as a list of predefined images}

\item{sigma}{standard deviation of the noise to add}

\item{smband}{smoothing band}

\item{tolmult}{multiplier of tolerance, default is set to 0.01, determines the cutoff for branching}

\item{margin}{space to leave between the demo image and boundary}

\item{slice_idx}{slices to plot}

\item{invert}{Invert image color}

\item{n_iter}{number of iterations}

\item{plot_instance}{how many instances to plot}
}
\value{
rmse of noisy and smoothed image, sample instances of noisy and smoothed
}
\description{
Title demo3diter
Multiple iterations of 3D demo
}
\examples{
demo3diter(image=128,sigma=0.1,smband=4,tolmult=0.01,margin=0.05,slice_idx=NULL,invert=FALSE,n_iter=10,plot_instance=1)
}
