% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothimage.R
\name{smoothimage}
\alias{smoothimage}
\title{Title smoothimage}
\usage{
smoothimage(img, minlossimprove = 1e-04, smband = 4)
}
\arguments{
\item{img}{Image to be denoised}

\item{minlossimprove}{cutoff for branching}

\item{smband}{smoothing band}
}
\value{
smoothed image, and decision tree
}
\description{
Title smoothimage
}
\examples{
smoothimage(img,0,0001,4)
}
